split Model and FrameworkModel tests into different files
